+++ server/configs/index.js
@@ -1,7 +1,7 @@
 module.exports = {
   database: {
-    host: '127.0.0.1',
-    port: '27017',
-    db_name: 'shop'
-  }
-}
+    host: "127.0.0.1",
+    port: "27017",
+    db_name: "orders",
+  },
+};
